+++ examples/reload-feature/lib/src/lib.rs
@@ -1,4 +1,4 @@
-#[unsafe(no_mangle)]
+#[cfg_attr(feature = "reload", unsafe(no_mangle))]
 pub fn do_stuff() {
     println!("step called");
 }