+++ macro/src/util.rs
@@ -1,6 +1,5 @@
+use proc_macro2::Span;
 use std::path::PathBuf;
-
-use proc_macro2::{Span, TokenTree};
 use syn::{Error, ForeignItemFn, LitStr, Result};
 
 pub fn ident_from_pat(
@@ -22,8 +21,9 @@ pub fn ident_from_pat(
 /// Reads the contents of a Rust source file and finds the top-level functions that have
 /// - visibility public
 /// - #[unsafe(no_mangle)] attribute
-///   It converts these functions into a [syn::ForeignItemFn] so that those can
-///   serve as lib function declarations of the lib reloader.
+///
+/// It converts these functions into a [syn::ForeignItemFn] so that those can
+/// serve as lib function declarations of the lib reloader.
 pub fn read_functions_from_file(
     file_name: LitStr,
     ignore_no_mangle: bool,
@@ -58,28 +58,68 @@ pub fn read_functions_from_file(
                 // we can optionally assume that the function will be unmangled
                 // by other means than a direct attribute
                 if !ignore_no_mangle {
-                    let no_mangle = fun
-                        .attrs
-                        .iter()
-                        .filter_map(|attr| attr.path().get_ident())
-                        .any(|ident| *ident == "no_mangle");
+                    fn cfg_no_mangle<'a>(
+                        mut cfg_items: impl Iterator<Item = &'a syn::Meta>,
+                    ) -> bool {
+                        let _predicate = cfg_items.next();
+                        // TODO: return false if predicate is false
+                        // false positives are unlikely, but can still compile error
+                        cfg_items.any(|meta| match meta {
+                            syn::Meta::Path(path) => path.is_ident("no_mangle"),
+                            syn::Meta::List(list) => {
+                                let mut found_no_mangle = false;
+                                if list
+                                    .parse_nested_meta(|meta| {
+                                        if meta.path.is_ident("no_mangle") {
+                                            found_no_mangle = true;
+                                        }
+                                        Ok(())
+                                    })
+                                    .is_err()
+                                {
+                                    return false;
+                                }
+                                found_no_mangle
+                            }
 
-                    let unsafe_no_mangle = fun.attrs.iter().any(|attr| {
-                        attr.meta.require_list().is_ok_and(|list| {
-                            list.path
-                                .get_ident()
-                                .is_some_and(|ident| *ident == "unsafe")
-                                && list.tokens.clone().into_iter().any(|token| {
-                                    if let TokenTree::Ident(ident) = token {
-                                        ident == "no_mangle"
-                                    } else {
-                                        false
-                                    }
-                                })
+                            _ => false,
                         })
-                    });
+                    }
 
-                    if !no_mangle && !unsafe_no_mangle {
+                    fn is_no_mangle<'a>(
+                        mut attrs: impl Iterator<Item = &'a syn::Attribute>,
+                    ) -> bool {
+                        attrs.any(|attr| {
+                            let ident = match attr.path().get_ident() {
+                                Some(i) => i,
+                                None => return false,
+                            };
+                            if *ident == "no_mangle" {
+                                true
+                            } else if *ident == "unsafe" {
+                                let mut found_no_mangle = false;
+                                if attr.parse_nested_meta(|meta| {
+                                    if meta.path.is_ident("no_mangle") {
+                                        found_no_mangle = true;
+                                    }
+                                    Ok(())
+                                }).is_err() {
+                                    return false;
+                                }
+                                found_no_mangle
+                            } else if *ident == "cfg_attr" {
+                                let nested = match attr.parse_args_with(syn::punctuated::Punctuated::<syn::Meta, syn::Token![,]>::parse_terminated) {
+                                    Ok(nested) => nested,
+                                    _ => return false,
+                                };
+                                cfg_no_mangle(nested.iter())
+                            } else {
+                                false
+                            }
+                        })
+                    }
+
+                    if !is_no_mangle(fun.attrs.iter()) {
                         continue;
                     };
                 }